Added new StreamStream class to move data between streams.
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@3236 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Sync/ClientFile.cs
@@ -417,7 +417,7 @@ namespace Simias.Sync
 				Stream inStream = response.GetResponseStream();
 				try
 				{
-					int bytesToWrite = Math.Min(sizeRemaining, (seg.EndBlock - seg.StartBlock + 1) * HashData.BlockSize);
+					int bytesToWrite = (int)Math.Min(sizeRemaining, (seg.EndBlock - seg.StartBlock + 1) * HashData.BlockSize);
 					Write(inStream, bytesToWrite);
 					sizeRemaining -= bytesToWrite;
 					eventPublisher.RaiseEvent(new FileSyncEventArgs(collection.ID, ObjectType.File, false, Name, fileSize, sizeToSync, sizeRemaining, Direction.Downloading));
